Refactored Mapper into individual modules
