+++ lib/mits.rb
@@ -2,22 +2,10 @@ require 'open-uri'
 require 'saxerator'
 require 'simple_objects'
 
-require 'mits/saxerator_ext'
-
 require 'mits/document'
 require 'mits/mapper'
+require 'mits/saxerator_ext'
 require 'mits/version'
-
-require 'mits/entities/address'
-require 'mits/entities/amenity'
-require 'mits/entities/company'
-require 'mits/entities/deposit'
-require 'mits/entities/fees'
-require 'mits/entities/file'
-require 'mits/entities/pet'
-require 'mits/entities/pet_policy'
-require 'mits/entities/property'
-require 'mits/entities/unit'
 
 module MITS
 end
